pyproject.toml
[tool.poetry]
name = "fortiproxy"
version = "0.1.0"
description = ""
authors = ["Kamil Podbielski <dev.sf4py@simplelogin.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyotp = "^2.7.0"
asyncclick = "^8.1.3.4"
pyppeteer = "^1.0.2"
anyio = "^3.6.1"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
